move SetupInfoUserMetrics out of a cgo build constraint
